Add mpv command
Added a new command to the bot that allows it to control an mpv instance.
The command can be triggered by sending the message !mpv <path-to-file>
to the bot. The implementation of the command includes using asyncio to
run subprocesses and a socket connection to communicate with the mpv
instance.